Fix audio not playing issue
The root cause was that dsp::sample_t was a double while the callback was expecting a float, so the out_buf_size was being computed with a value larger than the actual size and thus the memset in the AudioThreadState::PAUSED case was overflowing the buffer

+++ src/dsp_declarations.hpp
@@ -3,7 +3,7 @@
 #include <stdint.h>
 
 namespace dsp {
-using sample_t = double;
+using sample_t = float;
 using phase_t = int32_t;
 using frequency_t = double;
 using amplitude_t = double;
